refactor(crates/clipboard): rename `MockClipboard` to `LocalClipboard`
